requirements.txt
streamlit>=1.30.0
pycaret[full]
pandas
scikit-learn
python-dotenv
joblib
